# invoice-module/swagger.yaml
openapi: 3.0.3
info:
  title: API Module de Facturation
  description: |
    API REST pour la gestion des clients et factures.
    
    Cette API permet de :
    - Gérer les clients (CRUD)
    - Gérer les factures avec leurs lignes (CRUD)
    - Calculer automatiquement les montants HT, TVA et TTC
    - Générer des numéros de facture automatiquement
    - Obtenir des statistiques sur les factures
    
    ## Format de réponse standardisé
    
    Toutes les réponses suivent le format :
    ```json
    {
      "success": true|false,
      "data": { /* données */ },
      "message": "Message descriptif",
      "error": "Détails erreur (optionnel)",
      "errors": { /* erreurs validation (optionnel) */ }
    }
    ```
  version: 1.0.0
  contact:
    name: Support API
    email: support@facturation.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api
    description: Serveur de développement
  - url: https://api.facturation.com/api
    description: Serveur de production
paths:
...
  /clients:
    get:
      tags:
        - Clients
      summary: Liste tous les clients
      description: Récupère la liste complète des clients
      responses:
        '200':
          description: Liste des clients récupérée avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientsListResponse'
              example:
                success: true
                data:
                  - id: 1
                    nom: "Entreprise ABC"
                    email: "contact@abc.com"
                    siret: "12345678901234"
                    date_creation: "2024-01-15"
                    created_at: "2024-01-15T10:00:00.000000Z"
                    updated_at: "2024-01-15T10:00:00.000000Z"
                message: "Clients récupérés avec succès"
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      tags:
        - Clients
      summary: Crée un nouveau client
      description: Ajoute un nouveau client dans le système
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClientRequest'
            example:
              nom: "Entreprise ABC"
              email: "contact@abc.com"
              siret: "12345678901234"
              date_creation: "2024-01-15"
      responses:
        '201':
          description: Client créé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'
              example:
                success: true
                data:
                  id: 1
                  nom: "Entreprise ABC"
                  email: "contact@abc.com"
                  siret: "12345678901234"
                  date_creation: "2024-01-15"
                  created_at: "2024-01-15T10:00:00.000000Z"
                  updated_at: "2024-01-15T10:00:00.000000Z"
                message: "Client créé avec succès"
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServerError'
  /clients/{id}:
    get:
      tags:
        - Clients
      summary: Récupère un client spécifique
      description: Obtient les détails d'un client par son ID
      parameters:
        - $ref: '#/components/parameters/ClientId'
      responses:
        '200':
          description: Client récupéré avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'
        '404':
          $ref: '#/components/responses/ClientNotFound'
        '500':
          $ref: '#/components/responses/ServerError'

    put:
      tags:
        - Clients
      summary: Met à jour un client
      description: Modifie les informations d'un client existant
      parameters:
        - $ref: '#/components/parameters/ClientId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClientRequest'
            example:
              nom: "Entreprise ABC Modifiée"
              email: "nouveau@abc.com"
      responses:
        '200':
          description: Client mis à jour avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'
        '404':
          $ref: '#/components/responses/ClientNotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServerError'

    delete:
      tags:
        - Clients
      summary: Supprime un client
      description: Supprime un client du système (impossible s'il a des factures)
      parameters:
        - $ref: '#/components/parameters/ClientId'
      responses:
        '200':
          description: Client supprimé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                message: "Client supprimé avec succès"
        '400':
          description: Impossible de supprimer le client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "Impossible de supprimer le client car il a des factures associées"
        '404':
          $ref: '#/components/responses/ClientNotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /clients/{id}/invoices:
    get:
      tags:
        - Clients
      summary: Récupère les factures d'un client
      description: Obtient toutes les factures associées à un client
      parameters:
        - $ref: '#/components/parameters/ClientId'
      responses:
        '200':
          description: Factures du client récupérées avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoicesListResponse'
        '404':
          $ref: '#/components/responses/ClientNotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  parameters:
    ClientId:
      name: id
      in: path
      required: true
      description: ID unique du client
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 1
...
  schemas:
...
    Client:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID unique du client
          example: 1
        nom:
          type: string
          maxLength: 255
          description: Nom de l'entreprise cliente
          example: "Entreprise ABC"
        email:
          type: string
          format: email
          maxLength: 255
          description: Adresse email du client
          example: "contact@abc.com"
        siret:
          type: string
          pattern: '^[0-9]{14}$'
          nullable: true
          description: Numéro SIRET (14 chiffres)
          example: "12345678901234"
        date_creation:
          type: string
          format: date
          nullable: true
          description: Date de création du client
          example: "2024-01-15"
        created_at:
          type: string
          format: date-time
          description: Date de création de l'enregistrement
          example: "2024-01-15T10:00:00.000000Z"
        updated_at:
          type: string
          format: date-time
          description: Date de dernière modification
          example: "2024-01-15T10:00:00.000000Z"
      required:
        - id
        - nom
        - email
        - created_at
        - updated_at
    Invoice:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID unique de la facture
          example: 1
        client_id:
          type: integer
          format: int64
          description: ID du client associé
          example: 1
        numero_facture:
          type: string
          maxLength: 50
          description: Numéro unique de la facture
          example: "FAC-2024-0001"
        date_facture:
          type: string
          format: date
          description: Date d'émission de la facture
          example: "2024-01-20"
        total_ht:
          type: string
          format: decimal
          description: Montant total hors taxes
          example: "6600.00"
        total_tva:
          type: string
          format: decimal
          description: Montant total de la TVA
          example: "1320.00"
        total_ttc:
          type: string
          format: decimal
          description: Montant total toutes taxes comprises
          example: "7920.00"
        created_at:
          type: string
          format: date-time
          description: Date de création de l'enregistrement
          example: "2024-01-20T10:00:00.000000Z"
        updated_at:
          type: string
          format: date-time
          description: Date de dernière modification
          example: "2024-01-20T10:00:00.000000Z"
      required:
        - id
        - client_id
        - numero_facture
        - date_facture
        - total_ht
        - total_tva
        - total_ttc
        - created_at
        - updated_at
...
    CreateClientRequest:
      type: object
      properties:
        nom:
          type: string
          maxLength: 255
          description: Nom de l'entreprise cliente
          example: "Entreprise ABC"
        email:
          type: string
          format: email
          maxLength: 255
          description: Adresse email du client
          example: "contact@abc.com"
        siret:
          type: string
          pattern: '^[0-9]{14}$'
          nullable: true
          description: Numéro SIRET (14 chiffres)
          example: "12345678901234"
        date_creation:
          type: string
          format: date
          nullable: true
          description: Date de création du client
          example: "2024-01-15"
      required:
        - nom
        - email
    UpdateClientRequest:
      type: object
      properties:
        nom:
          type: string
          maxLength: 255
          description: Nom de l'entreprise cliente
          example: "Entreprise ABC Modifiée"
        email:
          type: string
          format: email
          maxLength: 255
          description: Adresse email du client
          example: "nouveau@abc.com"
        siret:
          type: string
          pattern: '^[0-9]{14}$'
          nullable: true
          description: Numéro SIRET (14 chiffres)
          example: "12345678901234"
        date_creation:
          type: string
          format: date
          nullable: true
          description: Date de création du client
          example: "2024-01-15"
...
    BaseResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indique si la requête a réussi
          example: true
        message:
          type: string
          description: Message descriptif
          example: "Opération réussie"
      required:
        - success
        - message
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            success:
              example: true
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            success:
              example: false
            error:
              type: string
              description: Détails techniques de l'erreur
              example: "Erreur technique détaillée"
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            errors:
              type: object
              description: Erreurs de validation par champ
              additionalProperties:
                type: array
                items:
                  type: string
              example:
                email:
                  - "Cette adresse email est déjà utilisée."
                siret:
                  - "Le SIRET doit contenir exactement 14 caractères."
    ClientResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Client'
    ClientsListResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Client'
...
    InvoicesListResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Invoice'
...
  responses:
    ClientNotFound:
      description: Client non trouvé
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Client non trouvé"
...
    ValidationError:
      description: Erreur de validation des données
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            success: false
            message: "Erreur de validation"
            errors:
              email:
                - "Cette adresse email est déjà utilisée."
              siret:
                - "Le SIRET doit contenir exactement 14 caractères."
    ServerError:
      description: Erreur interne du serveur
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Erreur interne du serveur"
            error: "Détails techniques de l'erreur"
...